Add - Timeline calls with queries and default queries for streaming posts
